Move enums to vocab.py; tweak inventory
